webhooks: Move github_webhook/ to github/.
