+++ mypy.ini
@@ -43,7 +43,7 @@ strict_optional = False
 strict_optional = False
 [mypy-zerver.webhooks.gitlab.view]
 strict_optional = False
-[mypy-zerver.webhooks.github_webhook.view]
+[mypy-zerver.webhooks.github.view]
 strict_optional = False
 [mypy-zerver.webhooks.bitbucket.view]
 strict_optional = False
