+++ templates/zerver/hello.html
@@ -522,7 +522,7 @@
                     <p class="integration-description">See build results immediately</p>
                 </div>
             </a>
-            <a href="/integrations/doc/github_webhook">
+            <a href="/integrations/doc/github">
                 <div class="group">
                     <img class="integration-logo" src="/static/images/integrations/logos/github.svg" alt="{{ _('Github logo') }}">
                     <h3 class="integration-name">Github</h3>
