+++ tools/test-backend
@@ -106,7 +106,7 @@ not_yet_fully_covered = {
     'zerver/webhooks/bitbucket2/view.py',
     'zerver/webhooks/freshdesk/view.py',
     'zerver/webhooks/github_legacy/view.py',
-    'zerver/webhooks/github_webhook/view.py',
+    'zerver/webhooks/github/view.py',
     'zerver/webhooks/gitlab/view.py',
     'zerver/webhooks/gogs/view.py',
     'zerver/webhooks/greenhouse/view.py',
