+++ zerver/webhooks/github/fixtures/delete.json
