+++ zerver/webhooks/github/tests.py
@@ -10,7 +10,7 @@ from zerver.models import Message
 class GithubWebhookTest(WebhookTestCase):
     STREAM_NAME = 'github'
     URL_TEMPLATE = "/api/v1/external/github?stream={stream}&api_key={api_key}"
-    FIXTURE_DIR_NAME = 'github_webhook'
+    FIXTURE_DIR_NAME = 'github'
     EXPECTED_SUBJECT_REPO_EVENTS = u"public-repo"
     EXPECTED_SUBJECT_ISSUE_EVENTS = u"public-repo / Issue #2 Spelling error in the README file"
     EXPECTED_SUBJECT_PR_EVENTS = u"public-repo / PR #1 Update the README with new information"
@@ -208,7 +208,7 @@ class GithubWebhookTest(WebhookTestCase):
             HTTP_X_GITHUB_EVENT='pull_request'
         )
 
-    @patch('zerver.webhooks.github_webhook.view.check_send_webhook_message')
+    @patch('zerver.webhooks.github.view.check_send_webhook_message')
     def test_pull_request_labeled_ignore(
             self, check_send_webhook_message_mock: MagicMock) -> None:
         payload = self.get_body('labeled_pull_request')
@@ -216,7 +216,7 @@ class GithubWebhookTest(WebhookTestCase):
         self.assertFalse(check_send_webhook_message_mock.called)
         self.assert_json_success(result)
 
-    @patch('zerver.webhooks.github_webhook.view.check_send_webhook_message')
+    @patch('zerver.webhooks.github.view.check_send_webhook_message')
     def test_pull_request_unlabeled_ignore(
             self, check_send_webhook_message_mock: MagicMock) -> None:
         payload = self.get_body('unlabeled_pull_request')
@@ -224,7 +224,7 @@ class GithubWebhookTest(WebhookTestCase):
         self.assertFalse(check_send_webhook_message_mock.called)
         self.assert_json_success(result)
 
-    @patch('zerver.webhooks.github_webhook.view.check_send_webhook_message')
+    @patch('zerver.webhooks.github.view.check_send_webhook_message')
     def test_pull_request_request_review_ignore(
             self, check_send_webhook_message_mock: MagicMock) -> None:
         payload = self.get_body('request_review_pull_request')
@@ -232,7 +232,7 @@ class GithubWebhookTest(WebhookTestCase):
         self.assertFalse(check_send_webhook_message_mock.called)
         self.assert_json_success(result)
 
-    @patch('zerver.webhooks.github_webhook.view.check_send_webhook_message')
+    @patch('zerver.webhooks.github.view.check_send_webhook_message')
     def test_pull_request_request_review_remove_ignore(
             self, check_send_webhook_message_mock: MagicMock) -> None:
         payload = self.get_body('request_review_removed_pull_request')
@@ -240,7 +240,7 @@ class GithubWebhookTest(WebhookTestCase):
         self.assertFalse(check_send_webhook_message_mock.called)
         self.assert_json_success(result)
 
-    @patch('zerver.webhooks.github_webhook.view.check_send_webhook_message')
+    @patch('zerver.webhooks.github.view.check_send_webhook_message')
     def test_push_1_commit_filtered_by_branches_ignore(
             self, check_send_webhook_message_mock: MagicMock) -> None:
         self.url = self.build_webhook_url(branches='master,development')
@@ -249,7 +249,7 @@ class GithubWebhookTest(WebhookTestCase):
         self.assertFalse(check_send_webhook_message_mock.called)
         self.assert_json_success(result)
 
-    @patch('zerver.webhooks.github_webhook.view.check_send_webhook_message')
+    @patch('zerver.webhooks.github.view.check_send_webhook_message')
     def test_push_50_commits_filtered_by_branches_ignore(
             self, check_send_webhook_message_mock: MagicMock) -> None:
         self.url = self.build_webhook_url(branches='master,development')
@@ -258,7 +258,7 @@ class GithubWebhookTest(WebhookTestCase):
         self.assertFalse(check_send_webhook_message_mock.called)
         self.assert_json_success(result)
 
-    @patch('zerver.webhooks.github_webhook.view.check_send_webhook_message')
+    @patch('zerver.webhooks.github.view.check_send_webhook_message')
     def test_push_multiple_comitters_filtered_by_branches_ignore(
             self, check_send_webhook_message_mock: MagicMock) -> None:
         self.url = self.build_webhook_url(branches='master,development')
@@ -267,7 +267,7 @@ class GithubWebhookTest(WebhookTestCase):
         self.assertFalse(check_send_webhook_message_mock.called)
         self.assert_json_success(result)
 
-    @patch('zerver.webhooks.github_webhook.view.check_send_webhook_message')
+    @patch('zerver.webhooks.github.view.check_send_webhook_message')
     def test_push_multiple_comitters_with_others_filtered_by_branches_ignore(
             self, check_send_webhook_message_mock: MagicMock) -> None:
         self.url = self.build_webhook_url(branches='master,development')
